Correction: Linus Torvald only wrote the GNU/Linux kernel
the GNU/Linux operating system is a collection of worldwide contributed software and drivers. Most of the developers came from the Unix world, the system was called after GNU == GNU is not UNIX.  Was not written by one single person. Their initial kernel had very bad performance, Linus Torvalds wrote a very performing kernel and in his honour they called the system GNU/Linux. The fact that everybody calls it just Linux nowadays is a sad fact of how history likes to cancel collective works and foster a single man superhero vision.  It is true that without Torvalds' kernel GNU/Linux would not have reached the quality it has today.